Adds a module `address_lookup_table::state` to the SDK.
