+++ core/src/banking_stage/consumer.rs
@@ -737,7 +737,7 @@ mod tests {
             unprocessed_transaction_storage::ThreadType,
         },
         crossbeam_channel::{unbounded, Receiver},
-        solana_address_lookup_table_program::state::{AddressLookupTable, LookupTableMeta},
+        solana_address_lookup_table_program::state::AddressLookupTable,
         solana_cost_model::cost_model::CostModel,
         solana_entry::entry::{next_entry, next_versioned_entry},
         solana_ledger::{
@@ -754,7 +754,7 @@ mod tests {
         solana_runtime::prioritization_fee_cache::PrioritizationFeeCache,
         solana_sdk::{
             account::AccountSharedData,
-            address_lookup_table,
+            address_lookup_table::{self, state::LookupTableMeta},
             instruction::InstructionError,
             message::{v0, v0::MessageAddressTableLookup, MessageHeader, VersionedMessage},
             poh_config::PohConfig,
