+++ programs/address-lookup-table-tests/tests/common.rs
@@ -1,13 +1,12 @@
 #![allow(dead_code)]
 use {
     solana_address_lookup_table_program::{
-        processor::process_instruction,
-        state::{AddressLookupTable, LookupTableMeta},
+        processor::process_instruction, state::AddressLookupTable,
     },
     solana_program_test::*,
     solana_sdk::{
         account::AccountSharedData,
-        address_lookup_table::program::id,
+        address_lookup_table::{program::id, state::LookupTableMeta},
         clock::Slot,
         hash::Hash,
         instruction::{Instruction, InstructionError},
