+++ programs/address-lookup-table-tests/tests/extend_lookup_table_ix.rs
@@ -3,11 +3,11 @@ use {
     common::{
         add_lookup_table_account, assert_ix_error, new_address_lookup_table, setup_test_context,
     },
-    solana_address_lookup_table_program::state::{AddressLookupTable, LookupTableMeta},
+    solana_address_lookup_table_program::state::AddressLookupTable,
     solana_program_test::*,
     solana_sdk::{
         account::{ReadableAccount, WritableAccount},
-        address_lookup_table::instruction::extend_lookup_table,
+        address_lookup_table::{instruction::extend_lookup_table, state::LookupTableMeta},
         clock::Clock,
         instruction::{Instruction, InstructionError},
         pubkey::{Pubkey, PUBKEY_BYTES},
