+++ programs/address-lookup-table/src/processor.rs
@@ -1,13 +1,14 @@
 use {
-    crate::state::{
-        AddressLookupTable, LookupTableMeta, LookupTableStatus, ProgramState,
-        LOOKUP_TABLE_MAX_ADDRESSES, LOOKUP_TABLE_META_SIZE,
-    },
+    crate::state::AddressLookupTable,
     solana_program_runtime::{declare_process_instruction, ic_msg, invoke_context::InvokeContext},
     solana_sdk::{
         address_lookup_table::{
             instruction::ProgramInstruction,
             program::{check_id, id},
+            state::{
+                LookupTableMeta, LookupTableStatus, ProgramState, LOOKUP_TABLE_MAX_ADDRESSES,
+                LOOKUP_TABLE_META_SIZE,
+            },
         },
         clock::Slot,
         feature_set,
